created file uploader and http requests to server

+++ backend/include/constants.h
@@ -4,5 +4,5 @@
 
 namespace Constants
 {
-    constexpr uint32_t dat_file_identifier = 0xDEADBEEF;
+    constexpr size_t dat_file_identifier = 0xDEADBEEF;
 }
